refactor : 디렉토리 구조 변경 및 컴포넌트명 변경
