+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import Calendar from "./calendar";
+import Calendar from "./components/calendar";
 
 function App() {
   return (
