+++ src/components/Button.tsx
@@ -1,0 +1,7 @@
+const Button = () => {
+  return (
+    <div>Button</div>
+  )
+}
+
+export default Button
